refactor: move app to own module so we can separate fn
